update: facemouse

+++ tools/img_to_webp.py
@@ -15,11 +15,14 @@ def convert_to_webp(target_dir: str, quality: int):
             if filename.lower().endswith(extensions):
                 img_path = os.path.join(root, filename)
                 output_path = os.path.splitext(img_path)[0] + ".webp"
-
+                if os.path.exists(output_path):
+                    do_overwrite = input(f"Do u want to overwrite '{output_path}'? (y/n): ")
+                    do_overwrite = do_overwrite.strip().lower() == 'y'
+                    if not do_overwrite:
+                        continue
                 try:
                     with Image.open(img_path) as img:
                         img.save(output_path, "webp", quality=quality)
-
                     os.remove(img_path)
                     print(f"Converted: {img_path} -> {output_path}")
                 except Exception as e:
